Updates design of header
