+++ src/app/app.component.css
@@ -22,7 +22,7 @@
 
 .menu {
     width: 216px;
-    background-color: rgba(255, 255, 255, 0.05);
+    background-color: rgba(255, 255, 255, 0.07);
     border: none;
 }
