--- src/app/components/account-actions/account-actions.component.css
@@ -1,5 +1,0 @@
-.account-button {
-    height: 47px !important;
-    width: 47px !important;
-    border: 2px solid #ffffff;
-}
