--- src/app/components/account-actions/account-actions.component.html
@@ -1,27 +1,0 @@
-<div fxLayout="row" fxLayoutGap="19px">
-    <a
-        *ngIf="(network$ | async)?.faucet"
-        [href]="faucetLink$ | async"
-        target="_blank"
-        mat-icon-button
-        class="account-button"
-        matTooltip="Open ether faucet"
-        fxLayout="column"
-        fxLayoutAlign="center center"
-        id="faucet-button"
-    >
-        <mat-icon svgIcon="faucet" aria-label="Open ether faucet"> </mat-icon>
-    </a>
-    <button
-        mat-icon-button
-        matTooltip="Show QR code to scan your address"
-        (click)="showOwnAddressQrCode()"
-        class="account-button"
-        fxLayout="column"
-        fxLayoutAlign="center center"
-        id="qr-button"
-    >
-        <mat-icon svgIcon="qr" aria-label="Show QR code to scan your address">
-        </mat-icon>
-    </button>
-</div>
