+++ src/app/components/address-input/address-input.component.html
@@ -23,8 +23,10 @@
                     trackBy: trackByFn
                 "
                 [value]="option.address"
+                class="contact-option"
             >
-                {{ option.label }} ({{ option.address }})
+                <span>{{ option.label }}</span
+                ><span>{{ option.address }}</span>
             </mat-option>
         </div>
     </mat-autocomplete>
