+++ src/app/components/contact-list/contact-list.component.css
@@ -1,5 +1,5 @@
 .list {
-    background-color: rgba(255, 255, 255, 0.05);
+    background-color: rgba(255, 255, 255, 0.06);
     border-radius: 35px;
     padding: 10px 20px 28px 20px;
 }
@@ -24,7 +24,7 @@
 .button {
     padding: 0 25px !important;
     color: #ffffff;
-    background-color: rgba(255, 255, 255, 0.05);
+    background-color: rgba(255, 255, 255, 0.06);
     height: 80px;
     z-index: 1;
 }
