+++ src/app/components/contact/contact.component.css
@@ -5,7 +5,7 @@
 .card {
     height: 80px;
     border-radius: 25px;
-    background-color: rgba(255, 255, 255, 0.05);
+    background-color: rgba(255, 255, 255, 0.06);
     padding: 0 25px;
     font-size: 13px;
     line-height: 15px;
