+++ src/app/components/search-field/search-field.component.css
@@ -1,19 +1,17 @@
 .search {
-    color: #000000;
+    color: rgba(255, 255, 255, 0.5);
     border-radius: 25px;
-    background-color: #393939;
+    background-color: rgba(255, 255, 255, 0.06);
     position: relative;
-    width: 297px;
 }
 
 .search__icon {
     position: absolute;
     left: 0;
-    margin-left: 18px;
-    height: 18px;
-    width: 18px;
-    line-height: 18px;
-    font-size: 18px;
+    margin-left: 16px;
+    height: 14px;
+    width: auto;
+    line-height: 14px;
 }
 
 .reset-button {
@@ -30,10 +28,9 @@
 }
 
 .search__input {
-    width: 198px;
+    width: 135px;
     height: 50px;
-    padding-left: 41px;
-    padding-right: 41px;
+    padding: 0 36px;
     background: none;
     border: none;
     box-shadow: none;
@@ -45,12 +42,20 @@
 
 .search__input::placeholder,
 .search__input::-webkit-input-placeholder {
-    color: #000000;
+    color: rgba(255, 255, 255, 0.5);
 }
 
 .search:focus-within,
 .search--focused {
+    color: #000000;
     background-color: #e7e7e7;
+}
+
+.search:focus-within .search__input::placeholder,
+.search:focus-within .search__input::-webkit-input-placeholder,
+.search--focused .search__input::placeholder,
+.search--focused .search__input::-webkit-input-placeholder {
+    color: #000000;
 }
 
 .options {
